chore: added prepare and distribute scripts
